requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
